requirements.txt
PyGithub==2.1.1
requests==2.31.0
python-dotenv==1.0.0
fastmcp==0.1.0
